+++ CHANGELOG.rst
@@ -1,6 +1,12 @@
 unreleased
 ----------
 
+- Remove add_user_attribute APIs from the agent.
+
+  The add_user_attribute and add_user_attributes APIs have been removed from
+  the agent.  These APIs have been replaced with
+  newrelic.agent.add_custom_parameter and newrelic.agent.add_custom_parameters.
+
 
 3.4.0 (2018-07-12)
 ------------------